refactor: Legacy MVC 삭제에 따른 테스트 코드 삭제 및 변경
