+++ app/src/main/java/com/techcourse/DispatcherServletInitializer.java
@@ -1,8 +1,11 @@
 package com.techcourse;
 
 import jakarta.servlet.ServletContext;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import com.interface21.DispatcherServlet;
 import com.interface21.web.WebApplicationInitializer;
 
 /**
@@ -17,7 +20,7 @@ public class DispatcherServletInitializer implements WebApplicationInitializer {
 
     @Override
     public void onStartup(final ServletContext servletContext) {
-        final var dispatcherServlet = new DispatcherServlet();
+        final var dispatcherServlet = new DispatcherServlet(Application.class.getPackage().getName());
 
         final var registration = servletContext.addServlet(DEFAULT_SERVLET_NAME, dispatcherServlet);
         if (registration == null) {
