--- app/src/test/java/samples/TestController.java
@@ -1,43 +1,0 @@
-package samples;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import com.interface21.context.stereotype.Controller;
-import com.interface21.web.bind.annotation.RequestMapping;
-import com.interface21.web.bind.annotation.RequestMethod;
-import com.interface21.webmvc.servlet.ModelAndView;
-import com.interface21.webmvc.servlet.view.JspView;
-
-@Controller
-public class TestController {
-
-    private static final Logger log = LoggerFactory.getLogger(TestController.class);
-
-    @RequestMapping(value = "/get-test", method = RequestMethod.GET)
-    public ModelAndView findUserId(final HttpServletRequest request, final HttpServletResponse response) {
-        log.info("test controller get method");
-        final var modelAndView = new ModelAndView(new JspView(""));
-        modelAndView.addObject("id", request.getAttribute("id"));
-        return modelAndView;
-    }
-
-    @RequestMapping(value = "/post-test", method = RequestMethod.POST)
-    public ModelAndView save(final HttpServletRequest request, final HttpServletResponse response) {
-        log.info("test controller post method");
-        final var modelAndView = new ModelAndView(new JspView(""));
-        modelAndView.addObject("id", request.getAttribute("id"));
-        return modelAndView;
-    }
-
-    @RequestMapping(value = "/all-test")
-    public ModelAndView all(final HttpServletRequest request, final HttpServletResponse response) {
-        log.info("test controller all method");
-        final var modelAndView = new ModelAndView(new JspView(""));
-        modelAndView.addObject("id", request.getAttribute("id"));
-        return modelAndView;
-    }
-}
