+++ mvc/src/test/java/com/interface21/webmvc/servlet/mvc/tobe/DispatcherServletTest.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet.mvc.tobe;
 
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.mock;
@@ -11,12 +11,13 @@ import jakarta.servlet.RequestDispatcher;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import jakarta.servlet.http.HttpSession;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
+
+import com.interface21.DispatcherServlet;
 
 public class DispatcherServletTest {
 
@@ -26,35 +27,10 @@ public class DispatcherServletTest {
 
     @BeforeEach
     void setUp() {
-        sut = new DispatcherServlet();
+        sut = new DispatcherServlet("samples");
         sut.init();
         request = mock(HttpServletRequest.class);
         response = mock(HttpServletResponse.class);
-    }
-
-    @Nested
-    @DisplayName("Controller 인터페이스 기반 컨트롤러에 대한 요청을 처리한다.")
-    class ControllerInterface {
-        @Test
-        @DisplayName("login에 대한 요청을 처리한다.")
-        void handleLogin() throws IOException, ServletException {
-            // given
-            final var session = mock(HttpSession.class);
-            final var requestDispatcher = mock(RequestDispatcher.class);
-
-            when(request.getParameter("account")).thenReturn("gugu");
-            when(request.getParameter("password")).thenReturn("password");
-            when(request.getRequestURI()).thenReturn("/login");
-            when(request.getMethod()).thenReturn("POST");
-            when(request.getSession()).thenReturn(session);
-            when(request.getRequestDispatcher(anyString())).thenReturn(requestDispatcher);
-
-            // when
-            sut.service(request, response);
-
-            // then
-            verify(response).sendRedirect("/index.jsp");
-        }
     }
 
     @Nested
@@ -67,7 +43,7 @@ public class DispatcherServletTest {
             // given
             final var requestDispatcher = mock(RequestDispatcher.class);
 
-            when(request.getRequestURI()).thenReturn("/register");
+            when(request.getRequestURI()).thenReturn("/get-test");
             when(request.getMethod()).thenReturn("GET");
             when(request.getRequestDispatcher(anyString())).thenReturn(requestDispatcher);
 
@@ -82,17 +58,18 @@ public class DispatcherServletTest {
         @DisplayName("/register에 대한 POST 요청을 처리한다.")
         void handleRegisterPOST() throws ServletException, IOException {
             // given
-            when(request.getParameter("account")).thenReturn("gugu");
-            when(request.getParameter("email")).thenReturn("gugu@email.com");
-            when(request.getParameter("password")).thenReturn("password");
-            when(request.getRequestURI()).thenReturn("/register");
+            final var requestDispatcher = mock(RequestDispatcher.class);
+
+            when(request.getParameter("id")).thenReturn("gugu");
+            when(request.getRequestURI()).thenReturn("/post-test");
             when(request.getMethod()).thenReturn("POST");
+            when(request.getRequestDispatcher(anyString())).thenReturn(requestDispatcher);
 
             // when
             sut.service(request, response);
 
             // then
-            verify(response).sendRedirect("/index.jsp"); // 리다이렉트 확인
+            verify(requestDispatcher).forward(request, response);
         }
     }
 }
