+++ mvc/src/test/java/com/interface21/webmvc/servlet/mvc/tobe/handlerAdapter/HandlerAdapterRegistryTest.java
@@ -1,16 +1,12 @@
 package com.interface21.webmvc.servlet.mvc.tobe.handlerAdapter;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.Mockito.mock;
-
-import java.util.NoSuchElementException;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
 import com.interface21.webmvc.servlet.mvc.tobe.handlerMapping.HandlerExecution;
 
 class HandlerAdapterRegistryTest {
@@ -31,15 +27,5 @@ class HandlerAdapterRegistryTest {
         var actual = sut.getHandlerAdapter(handler);
 
         assertThat(actual).isInstanceOf(AnnotationHandlerAdapter.class);
-    }
-
-    @Test
-    @DisplayName("처리 가능한 어댑터가 없는 경우 예외를 던진다.")
-    void getHandlerAdapter_null() {
-        var handler = mock(Controller.class);
-
-        assertThatThrownBy(() -> sut.getHandlerAdapter(handler))
-                .isInstanceOf(NoSuchElementException.class)
-                .hasMessage("No handlerAdapter found");
     }
 }
